requirements.txt
solara>=1.14.0
openai>=1.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
